DjangoRestAPI/requirements.txt
djangorestframework
markdown 
django-filter